requirements.txt
libtmux
docutils
pyperclip
pyyaml
pyfzf
textual
rich
